+++ src/main/java/com/project/csm/model/Movie.java
@@ -17,7 +17,7 @@ public class Movie {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "MovieID")
-	private int movieID;
+	private Long movieID;
 
 	@Column(name = "Name")
 	private String name;
